mv e2e assertions to sanity_checks.py
